// solace-quoteappdemo-frontend/src/app/model/stock-symbol.model.ts
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export interface StockSymbol {
  id: string;
  displayName: string;
  pOpen: number;
  pHigh?: number;
  pLow?: number;
  pClose: number;
  pDiff?: number;
  /**
   * Currently we need outside method to operate this value.
   */
  isRised?: boolean;
  /**
   * True if client subscribes this symbol, otherwise false.
   */
  isSubscribed?: boolean;
  vTotal?: number;
}